Renamed components to BreadcrumbTrail, StatefulBreadcrumbTrail
