+++ src/components/navbar/NavBar.tsx
@@ -20,7 +20,7 @@ import { useRouter } from "next/router";
 import * as React from "react";
 import { MouseEventHandler } from "react";
 import $navbar from "./Navbar.module.css";
-import BreadcrumbsNav from "../design_system/breadcrumbs/BreadcrumbsNav";
+import StatefulBreadcrumbTrail from "../breadcrumbs/StatefulBreadcrumbTrail";
 
 interface NavItemProps {
   href?: string;
@@ -115,7 +115,7 @@ export default function NavBar() {
                 {logo}
                 {drawer}
               </Box>
-              <BreadcrumbsNav />
+              <StatefulBreadcrumbTrail />
             </>
           ) : (
             <>
